.WAIT for makefs, because in most builds, it's the first target which includes Makefile.nbincludes, whose beforedepend: has some race condition I don't fully understand, and which has been causing many (but not all!) -j16 builds on the autobuild cluster to fail in tools since sometime between 201103061600 and 201103070900.
With the proper voodoo, I was able to get 8 builds in a row to fail
without this change, and 16 in a row to succeed with it.

XXX a proper fix likely involves reworking the beforedepend: target
of Makefile.nbincludes, but this should get the build cluster back up
to full functionality in the meanwhile.